A  program that takes in a string as input, detects if the string is an isogram and outputs true or false based on the result.
An isogram is a word that has no repeating letters, whether they are consecutive or non-consecutive.

Explanation on the code above: The first line of code is the function created using def .
The second line of code converts the word into lowercase to make the check case-insensitive. 
The word is converted into a set which removes duplicates letters. 
If the length of the set word is equal to the length of the word ,it is an isogram.

The input turbulence has repeated letters like 'u' and 'e' so it is not an isogram word .
The background word has unique set of word which make it an isogram.